+++ license_tools/tools/scancode_tools.py
@@ -310,7 +310,7 @@ class LicenseMatch:
     The detected license expression.
     """
 
-    spdx_license_expression: str
+    license_expression_spdx: str
     """
     The detected license expression in SPDX format.
     """